feat(centrifuge-js): expose static utils as named exports
