+++ centrifuge-js/src/index.ts
@@ -2,5 +2,6 @@ import { Centrifuge } from './Centrifuge'
 export { Collection, NFT } from './modules/nfts'
 export { Investment, Loan, LoanStatus, Pool, PoolRole } from './modules/pools'
 export * from './types'
+export * from './utils'
 
 export default Centrifuge
